% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_currency.R
\name{convert_currency}
\alias{convert_currency}
\title{Converts currency using api.exchangerate.host}
\usage{
convert_currency(from = "USD", to = "HUF", retried = 0)
}
\arguments{
\item{from}{Currency to convert from}

\item{to}{Currency to convert to}

\item{retried}{Hidden parameter used as an exponent of wait time between failed attempts.}
}
\value{
exchange rate
}
\description{
Converts currency using api.exchangerate.host
}
\examples{
convert_currency(from = "EUR", to = "HUF")
}
