% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forint.R
\name{forint}
\alias{forint}
\title{Formats numbers as Hungarian Forints}
\usage{
forint(x)
}
\arguments{
\item{x}{A number}
}
\value{
string
}
\description{
Formats numbers as Hungarian Forints
}
\examples{
forint(42)
forint(1231232351.323)
}
