% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_currency.R
\name{get_exchange_rates}
\alias{get_exchange_rates}
\title{Look up the historical value of a symbol in a specified base currency in a given time period.}
\usage{
get_exchange_rates(
  symbol = "HUF",
  base = "USD",
  start_date = Sys.Date() - 30,
  end_date = Sys.Date(),
  retried = 0
)
}
\arguments{
\item{symbol}{Currency to convert to}

\item{base}{Currency to convert from}

\item{start_date}{date}

\item{end_date}{date}

\item{retried}{Hidden parameter used as an exponent of wait time between failed attempts.}
}
\value{
\code{data.table} object
}
\description{
Look up the historical value of a symbol in a specified base currency in a given time period.
}
