% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_currency.R
\name{get_usdhufs}
\alias{get_usdhufs}
\title{Look up the historical value of a US Dollar in Hungarian Forints}
\usage{
get_usdhufs(start_date = Sys.Date() - 30, end_date = Sys.Date(), retried = 0)
}
\arguments{
\item{start_date}{date}

\item{end_date}{date}

\item{retried}{Hidden parameter used as an exponent of wait time between failed attempts.}
}
\value{
\code{data.table} object
}
\description{
Look up the historical value of a US Dollar in Hungarian Forints
}
